Add files via upload
This file is the updated file using Tai's grid code.
What I've added is that the code counts the number of people in each status and those numbers can be appended to the graph code.
Furthermore, I've updated the code so that it takes into account vaccinations. From the beginning of the code people can be vaccinated after a few days into the code.